Merge in server component.
History is squashed together, I know that history is super valuable to
have, but it's very time consuming to go over the long history of the
project and make sure that no commits/messages contain anything
private/confidential/referencing a specific customer (as much as we
tried to ensure that during development).

In addition, this is not a direct export of the Svix server, but rather
an adjustment of the internal code-base to make it more suitable to
being an open source project. For example, we added more configurability
here, removed parts that don't make sense, and removed a lot of the
operational dependencies that would make self-hosting it annoying.