Db models: use the same timestamp for ID and created_at. (#355)
We want them to be the same because they actually hold the same
information. We show the created_at, but we use the ID for the sorting.

We should probably get rid of created_at down the line because it's
already encoded in the ID, but at least then, make sure they are
identical.