Remove the multi-step Rust check slow and not needed
The idea was to do a quick check before doing the full check across
different versions. The other idea was to only run clippy an rustfmt
once.

Though since we want clippy and rustfmt to run on both versions anyway
(in case those changed), and that it's actually quick to do so, it made
more sense to just run both on both without the multi-step.