Python: fix issue with installing package on windows.
The problem was that we weren't explicitly setting the charset/encoding
when loading the README and it's different on every platform.
While the default is utf-8 (I think) on Linux and macOS, it's cp1252 on
Windows. The latter wasn't able to load the README which contains emoji
and other fancy stuff.

This fixes it to use utf-8 explicitly on every platform (like we already
did for another file we load here).