Change to a more efficient first message dispatch (#425)
Introduces the queue type `MessageBatch` for incoming messages so
we can concurrently send a message to multiple endpoints from within
`worker.rs`. The current queue type `MessageTask` is kept intact and
still used for message retries and other scenarios where sending to
a single endpoint is the desired behavior.

Note: adds a dependency on the `futures` crate.