Add soft and hard caps to pagination limits (#445)
Adds soft and hard caps to the limit for pagination. This is configured via a constant in the v1::utils module.

With soft caps no error is returned exceeding the cap, but list responses will use the cap instead of an exceeding value. With hard caps a 422 error is returned if the given limit exceeds the cap.