Implement Dashboard Access Endpoint (#488)
Implements the /auth/dashboard-access/:app_id/ endpoint and associated permissions changes. This makes it so there are tokens valid only for some endpoints. This is meant so that someone using the Svix service may create a token for the owner of an application under their organization to perform limited operations on what falls under the scope of that application.