Cleanup OpenAPI generation and print a note about Python (#494)
We had two different ways to run the OpenAPI generation. Both the
package.json and the regen_openapi.sh script.
And we weren't generating python correctly in neither of them (so added
a print for that).

This PR fixes it so that the generation is done correctly and only in
one place.