Python: update httpx and openapi-python-client. (#502)
httpx had a CVE: CVE-2021-41945
We were not affected by it because it uses code paths that we don't use,
though being a dependency we were still forcing an old version which is
not good.

We would have updated it before, but openapi-python-client was blocking
us. They just relesaed a version.