Server: fix caching dependencies for docker builds
The problem is that in order to build deps we need the full source tree.
This fix just adds a dummy source tree so that we can build and cache
the deps before doing a full build.

This is still an incomplete solution. It caches all of the deps, but it
still doesn't cache the building of the main binary. So it's 10x better
than what we have now, but there's still 10x more to go.

One day Cargo will hopefully fix it:
https://github.com/rust-lang/cargo/issues/2644

Solution from:
https://stackoverflow.com/questions/58473606/cache-rust-dependencies-with-docker-build