Server: make the application portal URL configurable.
This shouldn't be used for now, and is only used for internal purposes.
In the future we may want to expose it in the configuration to make it
easy to implement custom application portals.

We just don't want to do it yet to not force us to have the URL structure
a de-facto API.