Server: use correct timestamp for attempt's id and created (#517)
We were generating a timestamp at the beginning of the function which we
were using for generating the header timestamp. Though we weren't using
this timestamp for neither the Ksuid nor the created_at of attempt.
Causing them to have different values than what we expect them to be.

This is a fix for this occurrence, though the real problem is that we
are quite fragile about this. We should have a good way for generating
it that is not as easy to get wrong as this one.

I think the REAL problem though, is that we have the timestamp in two
places, the ksuid and the created_at field. Which we should definitely
fix.