Implement basic wait-for functionality (#484)
On forming a connection to PostgreSQL or Redis and failing, the server will no
longer immediately panic as long as the --wait-for flag is passed. Instead it
will loop forming a connection until the configured timeout period (set via
--wait-for <duration in seconds>) has been reached.

This has been tested manually and works as described. Programmatic testing is
tricky with this PR, so it is not included.

Fixes #334