Server CI: changed --frozen to --locked due to git dep (#526)
Because of the git dependency that we have, we need --locked to be
passed, not --frozen. Because we do want to let it access the internet
to check the git repo is fine, but still not actually make any
modifications to the lock file.

This fixes the build in some scenarios.