Server: add support for encrypting webhook secrets
Webhook secrets are used for signing the webhook messages and they are
stored in the main database and cache. Normally they are stored there in
the clear, which means that people with access to the database could
potentially retrieve them and forge webhook messages.

With this change, we are adding support for an optional secret key that
can be used to encrypt the webhook secrets.

This change is fully backwards compatible when encryption is turned off,
and you can also change from non-encrypted to encrypted whenever you
want and everything will continue to work.

Changing the encryption key, or moving from encrypted to non-encrypted
is not supported, as without the key you won't be able to view endpoint
secrets. We could potentially add a script that migrates the database
from one to the other in the future if needed, but it's probably a very
rare requirement.