Server: update tower-http and fix CORS.
We updated tower-http which then complained about bad CORS. So we fixed
the CORS settings and also added the missing CORS cache directive
(max_age).

The main issue was mixing wildcard origin and allow_creds.
It's not actually needed, because we can just explicitly allow the
`Authorization` header which is a better idea anyway because we don't
need creds as we don't support cookie authentication.

https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Origin

Thanks to @svix-daniel for finding this issue and figuring it out.