Add optional prefix parameter to kv_def macros and use where relevant (#568)
While the kv_def macro was there previously, we did not use it in some places
because of the need for custom prefixes. This PR augments the kv_def and
string_kv_def macros to take an optional string literal to be used as the
PREFIX_CACHE constant for the CacheKey trait.

Additionally this replaces the two manual implementations of the cache traits
with this new macro syntax.