Eliminate period in which data loss could occur rarely during crashes (#569)
Prior to these changes, there was an error where pending tasks being reinserted
into the queue were ACKed before the reinsertion meaning an improperly timed
crash could lead to lost message tasks. These changes reorder the ACKing and
reinsertion and perform them in a Redis pipeline such as to reduce the chance of
data loss on a server crash.