Introduce svix.NullableString utility (#589)
The MessageIn struct has an EventId field of type openapi.NullableString. The openapi.NullableString is defined in an internal package, so it's impossible to be imported by the end developer.

This PR introduces the svix.NullableString public utility to generate an openapi.NullableString struct to be used as an EventId.