JavaScript: change the utf8 encoder to the default node one
The TextEncoder API is supported both by Node and by web APIs, so we
should have used it in the first place though we didn't know it was
incomplete.

Essentially the problem is that we used a lib that doesn't support the
full extended utf8 (6 byte) range so it was barfing.

Here's a utf8 sequence that triggers the use (when encoded as string):

```
Uint8Array(21) [
  226, 143, 176, 240, 159,
  154, 180, 240, 159, 143,
  187, 226, 128, 141, 226,
  153, 130, 239, 184, 143,
   32
]
```

Initially reported on Slack:
https://svixcommunity.slack.com/archives/C022BJ2EBHV/p1659020835781289