Upsert behavior on PUT methods (#605)
This PR enables support for creation of entities during PUT requests
if the provided ID does not match an existing record. Affected
entities are Application, Endpoint, and Event Type. PUT requests
against those entities will now return 200 to indicate the record
existed and was modified, or 201 to indicate the record was created.

Closes #603.