Server: add support for P256 signature schemes.
This is similar to the asymmetric signatures added in 24bf8e5d740ec6d6f3a878a412c736e19c7da156
though those were using ed25519, and this one is using NIST P-256 ECDSA.

This is useful for people who care about NIST for  FIPS 140 certification
or other compliance reasons.

This is fully interoperable with the ed25519 and symmetric modes, and
can be switched live. Signatures use the `v1b` signature tag, and keys
use the `whsk1_` and `whpk1_` prefixes.