Server: include both DER and normal signatures for P256
We were only using the normal (r||s) serialized bytes signatures for
P256. While this was good in most cases, OpenSSL, some libraries, and
AWS make it hard to work with non-DER encoded stuff.

We now include both so it's easy for people using both OpenSSL style and
the "normal" style to verify signatures.