Ensure Cargo.lock is up to date with Cargo.toml (#667)
Recent Docker releases failed because of a frozen `Cargo.lock` on
release. Upon a little investigation it was found that the `Cargo.lock`
was slightly out of sync with the `Cargo.toml`. This change will
probably fix the Docker releases by ensuring the `Cargo.lock` matches
what happens when the current `main` branch is built normally.