Ensure QueueTasks are deleted after acknowledgement (#671)
Before this change, it was discovered that `QueueTasks` were only ACKed,
not deleted from the queue after dispatch. While this did not alter
functionality at all, it would lead to infinitely increasing memory use
with each `QueueTask`.

This was due to a confusing behavior in Redis, where an `XACK` would
remove the item from the list of items returned, but not actually free
the memory.