Server: omit prevIterator when null. (#715)
We use the ListResponse type everywhere where we have lists, and we
haven't added prev iterator support for all of them. This means that for
those we haven't supported, we were just seeing prevIterator=null which
is confusing and wrong.

This change omits the serialization of prev iterator when it's none. At
least until we actually get around to implementing it for all of the
routes.