Worker: fix small payload serialization for signatures
There was an issue with small payloads (well potentially larger too, but
exceedingly unlikely) which could cause the payloads to not serialize
correctly which could cause signatures to be incorrect in same rare
scenarios likely only affecting small payloads.

If memory serves, serde parses and stringifies json in a
non-deterministic way. Normally it would be deterministic, but it also
maintains an internal cache of "json structures", and if it thinks that
it can reuse one it does. This means that if there's a "cache hit",
which I'm not sure about the circumstances, it can reuse one of those
and therefore generate a different string.

Why does it matter? Because instead of using the string representation
of the payload, we were passing a `Json::Value` around and stringifying
and parsing it in multiple places. This is both inefficient, and
incorrect. Cryptographic hashes (like the ones used by the Svix
signature scheme) are very sensitive to even the slightest changes, not
to mention reordering, so we must use the same payload everywhere. We
can't just hope it'll be the same when strinigifying later.