Server: improve startup logs to indicate status
We were only logging about initialization starting, and never about the
success status. This meant that it wasn't very clear for someone
reviewing the logs what is going on exactly.

We've cleaned up the logs to be more consistent, and print for both
start of an operation and the end (success).