Server: improve startup logs to indicate status (#769)
We were only logging about initialization starting, and never about the
success status. This meant that it wasn't very clear for someone
reviewing the logs what is going on exactly.

We've cleaned up the logs to be more consistent, and print for both
start of an operation and the end (success).

Reason for the change: this just confused a customer (and me!) when we
were trying to figure out what's going on. The issue was with the server
not connecting to redis, but it wasn't obvious from the logs.