Add new endpoint for app portal access with a body
PR #765 added support for attaching feature flags to auth tokens. A
request body requirement was added to the dashboard access endpoint, but
this makes it incompatible with existing client libraries because they
do not send a request body.

This change essentially revets that by changing `/dashboard-access/`
back to its old behaviour of not accepting a body, and adds a new
endpoint, `/app-portal-access/`, which does.