Fix 5xx responses during concurrent UID updates (#848)
Followup to #846 which identifies and fixes the same problem in other
parts of the code. In brief, with unluckily timed concurrent PUT/PATCH
requests to the same object, the checks for UID conflicts done during
the start of the request would become stale as one request processed.
This leads to a specific `DbErr` which the changes now manually map to a
HTTP CONFLICT response.