Centralize the handling of deleted flags in CreateMessageApp (#854)
It was discovered that deleted applications could still be improperly
handled if a `CreateMessageApp` 's deleted fields (either on the
`struct` itself or via the `endpoints` field which each could be
deleted) was not respected. These changes ensure that deleted
application models return `Ok(None)` instead of `Some`
`CreateMessageApp`, centralizing the logic to the only constructor of
the `struct`.

Additionally, to prevent such issues from ever occurring the
`CreateMessageApp` has been changed slightly such that the `endpoints`
and `deleted` fields are private. This means that the struct must be
always be constructed via the `layered_fetch` which now filters for
deleted applications. And it means that endpoints must be fetched
through the `filtered_endpoints` method, which filters disabled and
deleted endpoints from the `Vec` of associated `CreateMessageEndpoint`s.

Finally, these changes introduce two tests ensuring the soundness of
this logic.

Do note that the `CreateMessageApp` is cached for 30 seconds, so it will
take at most that long for deletion of applications or endpoints to be
respected. This is bypassed in the aforementioned tests by manually
deleting the key for the `CreateMessageApp`.