Create message: change to RawValue instead of Value
We were using the `serde_json::Value` for the message payload which resulted
in a lot of allocations when the payload was complex, and was causing high
memory usage and fragmentation. With this, we don't actually parse this part
of the json, but rather just treat it as a string (serde verifies it's valid
json though). We unfortunately turn it into a `Value` when persisting to
PG, we should fix this extra thing that diminishes the point of this
change a bit.