add js transforms to bridge (#917)
Embeds a deno runtime in the bridge binary to allow ad hoc reshaping of
messages read from queues before they are sent to Svix.

## Motivation

When consuming JSON from messaging systems to generate new webhooks,
it's common to need to transform the payload before making the Create
Message request to Svix.

## Solution

Embedding a JS runtime into the bridge process allows us to run
user-defined functions on the payloads as they move from input to
output.

With this diff, plugin instances provided by
`svix-webhook-bridge-plugin-queue-consumer` can now evaluate js included
in the plugin config:
```yaml
plugins:
- type: "rabbitmqconsumer"
  input:
    uri: "amqp://guest:guest@localhost:5672/%2f"
    queue_name: "local"
    requeue_on_nack: false
  transformation: | 
    function handler(input) {
      return { 
        app_id: input.key, 
        message: { 
          eventType: input.event_type, 
          payload: input.data 
        }
      };
    }
  output:
    token: "***************"
```

The `transformation` key can now be set on any of these consumers. The
JS fragment should include a default export of a function that accepts
an object and returns an object.

These functions can be used to reshape the payload as necessary.