fix deno failure in bridge docker build
Looks like `deno_core` depends on `rusty_v8` which will try to download
prebuilt a binary v8 archive via some python script, and when that fails
it'll try again with `curl`.

The python download was failing. Why? No idea. The curl fallback also
failed because curl was not installed.

This diff adds `curl` to the list of installed packages for the builder
stage of the `Dockerfile` which allows the `cargo build` layer to
complete successfully.