polish bridge config (#925)
Alright y'all. After a bunch of conversations about the overall config
and mental model to apply to the project (from the user perspective),
this PR ended up _huge_. It'll be a tough review. I'm sorry :sob:

This is a pretty large shift.

Here's a breakdown of the broad strokes:

- The `Plugin` trait was removed from `-types`.
- New traits added: `SenderInput` and `ReceiverOutput` to represent the
  non-svix portions of the dataflow.
- Removed the `-plugin-webhook-receiver` plugin crate.
- moved the http server from the `-plugin-webhook-receiver` crate into
`svix-bridge`.
  - "queue forwarders" into the `-plugin-queue-consumer` crate.
- Renamed the `-plugin-queue-consumer` crate to just `-plugin-queue` (it
now provides both `SenderInput` and `ReceiverOutput` implementation for
the various queue backends.
- all crates etc have been renamed: `s/svix-webhook-bridge/svix-bridge/`


What else, what else, what else...

The example config has been split into 2: one showing `senders` and
another for `receivers`.
The readme has been gutted (but probably needs to be fleshed out a bit
more... again). It felt silly to effectively duplicate the example
configs in code fence blocks a 2nd time.

To sweeten the deal, new test coverage has been added for select
portions of the config parsing as well as for the HTTP receiver stuff.

Both the sender and receiver example configs are now checked in the test
suite to make sure they at least parse without failure. These tests
should be expanded to inspect the actual data, but this diff is so
big... :sob:

We didn't have any tests for the HTTP handler execution before this, so
that's very nice to have.