rename bridge ci
At one point bridge was being called "webhook bridge" and during the
renaming some references to the old name were missed.

Most of these omissions are harmless but in the case of the CI workflow
name, it mismatched with the updated name used in the badge on the
readme.

This should fix the broken link for the badge.