linter fixups for java lib (#956)
This is just to quiet the linter, but it's unclear why the linter woke
up all of a sudden, but this seems to sate it.

<!--
Thank you for your Pull Request. Please provide a description above and
review
the requirements below.

Bug fixes and new features should include tests.
-->

## Motivation

<!--
Explain the context and why you're making that change. What is the
problem
you're trying to solve? If a new feature is being added, describe the
intended
use case that feature fulfills.
-->

## Solution

<!--
Summarize the solution and provide any necessary context needed to
understand
the code change.
-->