Bridge: receivers now publish to queues with `payload` field
This diff is a little more broad than just "put the payload in a field."
It also tries to formalize the requirements of each output case, both
Senders and Receivers.

For Senders, we need a certain shape useful for building a call to
Create Message.
For Receivers, outputs expect to get a JSON object with a `payload`
field on it, representing the data to publish to the queue or what have
you.

These boundary types are now both in the `-types` crate, and used to
ensure both as-is and transformed data conforms where needed.

Existing tests caught the changes on the receiver side (which is
the main thing for this diff).

Additionally, for Senders, I camelCased the `appId` field to make the
create message JSON cased more consistently. This was a remark that came
up during review of an up-coming blog post.