Bridge: receivers with no transformation set body as payload field
Follow-on from #982.

For cases where the webhook receiver has not transformation set, we
assume the body is JSON. Formerly, that JSON body would fail to validate
if it didn't happen to have a root key named "payload" to align with our
expectations. This is a bad default.

Instead, we take the body, parse it as JSON, then shove it in the
payload field.