Bridge: variable expansion for config files
Config files can now use "dollar brace" notation to do variable
substitutions. See the readme for an example.

Furthermore, the variables available for substitution are exposed via a
simple HashMap meaning we can filter the vars exposed in the future, if
requested. This also has the benefit of being nicer for testing ;)

Tests are added to capture the initial behavior of the new dep, but also
to lock in the contract if we decided to "roll our own" in the future.

---

Open source review required.

Adds a dependency on [envsubst-rs](https://github.com/coreos/envsubst-rs)
to provide the base substitution capability.

This was selected based on the following criteria:

- support for the absolute minimum features (no expression evaluation
  for fallbacks, etc).
- Tiny. 170LOC (60+ of which are tests), easy to understand and fork if
  needed.
- Vaguely looks maintained, owned by coreos (part of redhat).

The big idea with adding this dep is, _"it already exists, but rolling
our own based on this if we want to customize behavior is easy."_