Bridge: load config from string (env var or cli flag)
Allows Bridge to load its config data from a string, which can be
supplied as an env var or directly on the cli with `--cfg`.

The original flag for specifying an alternative file path to read config
data from (formerly `--cfg` or `-c`) has been renamed to `--cfg-file`.

Sorry that's so confusing!

- Use `--cfg` to pass config as string
- Use `--cfg-file` to read from file, non-default loction

Additionally, a naive "search path" has been provided to look for the
default config file using a series of names:

- `svix-bridge.yaml`
- `svix-bridge.yml`
- `svix-bridge.json`

This was done specifically to formalize the fact we're advertising "we
accept json config data" in the readme.